+++ src/Factory/MigratorFactory.php
@@ -5,12 +5,16 @@ declare(strict_types=1);
 namespace Sirix\Cycle\Factory;
 
 use Cycle\Database\Exception\ConfigException;
-use Cycle\Migrations;
+use Cycle\Migrations\Config\MigrationConfig;
+use Cycle\Migrations\FileRepository;
+use Cycle\Migrations\Migrator;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
+use Sirix\Cycle\Internal\MigrationsToggle;
 use Sirix\Cycle\Service\MigratorInterface;
 use Sirix\Cycle\Service\MigratorWrapper;
+use Sirix\Cycle\Service\NullMigrator;
 
 final class MigratorFactory
 {
@@ -20,6 +24,10 @@ final class MigratorFactory
      */
     public function __invoke(ContainerInterface $container): MigratorInterface
     {
+        if (! MigrationsToggle::areMigrationsEnabled()) {
+            return new NullMigrator();
+        }
+
         $config = $container->has('config') ? $container->get('config') : [];
 
         if (! isset($config['cycle']['migrator'])) {
@@ -28,14 +36,14 @@ final class MigratorFactory
 
         $config = $config['cycle']['migrator'];
 
-        $migratorConfig = new Migrations\Config\MigrationConfig($config);
+        $migratorConfig = new MigrationConfig($config);
 
         $dbal = $container->get('dbal');
 
-        $migrator = new Migrations\Migrator(
+        $migrator = new Migrator(
             $migratorConfig,
             $dbal,
-            new Migrations\FileRepository($migratorConfig)
+            new FileRepository($migratorConfig)
         );
 
         return new MigratorWrapper($migrator);